+++ apps/web/src/app/[locale]/(admin)/new/loading.tsx
